requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
gevent>=1.4
boto3==1.34.45
botocore==1.34.45
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.10.0
Cython==3.0.8
defusedxml==0.7.1
Deprecated==1.2.14
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.60.1
gunicorn==21.2.0
h5py==3.11.0
huggingface-hub==0.20.3
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
keras==3.2.1
kiwisolver==1.4.5
libclang==16.0.6
line-bot-sdk==3.9.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.19
rsa==4.9
s3transfer==0.10.0
sahi==0.11.15
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
supervision==0.18.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torchvision==0.17.0
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.16
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
yolov5==7.0.13
